Add input to filter workers
