+++ client/src/pages/ShowWorkers.js
@@ -1,20 +1,14 @@
 import React from "react";
-import PropTypes from "prop-types";
 import { Container } from "reactstrap";
-import { ShowWorkersTable, AddNewButton, Title } from "../components";
+import { WorkersData, Title } from "../components";
 
-const ShowWorkers = ({ trigger }) => {
+const ShowWorkers = () => {
 	return (
 		<Container>
 			<Title text="Cleaners" />
-			<AddNewButton pathname="/add-worker" />
-			<ShowWorkersTable trigger={trigger} />
+			<WorkersData />
 		</Container>
 	);
-};
-
-ShowWorkers.propTypes = {
-	trigger: PropTypes.bool,
 };
 
 export default ShowWorkers;
